# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/register": {
      "post": {
        "operationId": "UserController_registerUser",
        "summary": "Register a new user profile with default settings",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User & Settings"
        ]
      }
    },
    "/api/user/me": {
      "get": {
        "operationId": "UserController_getProfile",
        "summary": "Get current user profile and settings",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User & Settings"
        ]
      },
      "patch": {
        "operationId": "UserController_updateProfile",
        "summary": "Update user profile (name, email)",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User & Settings"
        ]
      }
    },
    "/api/user/settings": {
      "get": {
        "operationId": "UserController_getSettings",
        "summary": "Get user configuration settings",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User & Settings"
        ]
      },
      "patch": {
        "operationId": "UserController_updateSettings",
        "summary": "Update user configuration settings",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User & Settings"
        ]
      }
    }
  },
  "info": {
    "title": "QuickPlan API Spec",
    "description": "Voice-first Smart Todo + Reminder + Expense/IOU Engine API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "x-user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User full name",
            "example": "Abhi Sharma"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "abhi@example.com"
          }
        }
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateSettingsDto": {
        "type": "object",
        "properties": {
          "inputLanguage": {
            "type": "string",
            "description": "Preferred input language: AUTO, HINDI, ENGLISH",
            "example": "AUTO"
          },
          "outputLanguage": {
            "type": "string",
            "description": "Preferred output language: SAME, HINDI, ENGLISH",
            "example": "SAME"
          },
          "notificationsEnabled": {
            "type": "boolean",
            "description": "Notification master toggle flag",
            "example": true
          },
          "defaultReminderOffsetMinutes": {
            "type": "number",
            "description": "Default reminder offset minutes",
            "example": 15
          },
          "defaultPriority": {
            "type": "string",
            "description": "Default task priority: LOW, MEDIUM, HIGH, URGENT",
            "example": "MEDIUM"
          },
          "defaultCategory": {
            "type": "string",
            "description": "Default task category",
            "example": "General"
          },
          "currency": {
            "type": "string",
            "description": "Default currency symbol/code",
            "example": "INR"
          }
        }
      }
    }
  }
}